Add some simple batching fixes. Draw calls on GH from 146 -> 14.
(Relies on the unmerged PR to switch glyph rendering to RGBA8 textures for those numbers).

One of the major issues is that sometimes the batcher ping pongs between
textures when using the dummy white texture. This was previously RGB8
format, so any glyphs would break batching. It's now RGBA8 which means that
batching works well with glyphs. The proper fix will be to add a dummy white
image in both RGB8 and RGBA8 format - and make the batcher aware of this. It
can then select the dummy white image that will avoid breaking the batch.

Additionally, the AABB tree split threshold of 512 was resulting in nodes that were
around 256 pixels high on GH. This produces too many false overlap detections that
currently cause batch breaks. Switching to 1024 didn't seem to affect parallelism
of batch building on any sites I tested. The proper fix for this involves making
the overlap detection a bit smarter, but more importantly, detecting when overlaps
involve opaque items and avoiding batch breaks in these situations.

Fixes #35.