Add clipping for border edges.
This could be made significantly faster for some cases (e.g. for dashed borders
detect if any clipping is needed for the bounding rect and run through a fast path
in that common case). On the other hand, perhaps dashed borders etc are rare
enough that it doesn't really matter.