Use OpenGL texture arrays where available to improve batching.
This takes an upgrade to OpenGL 3.x on the Mac. Because the Mac OpenGL
implementation does not support the Compatibility Profile as far as I'm
aware, the shaders have been rewritten to use some preprocessor magic to
work on both OpenGL ES 2.1 and OpenGL 3.x.

Memory usage will likely be somewhat suboptimal until we are dynamically
resizing texture arrays.

Closes #38.