Add support for scrolling in iframes.
In addition to handling iframe scrolling, this changes items to be in local space relative to their scroll root.

This is very much a "work in progress" commit. There's *lots* of things that don't work:
 * Nested transforms + scroll layers
 * Render target effects + scroll layers
 * Render target effects with transforms.

However, the changes in this PR are at a stage where all (previously passing) reftests still pass, and it lays
the groundwork for supporting those features listed above.