Expand viewports to consist of a rectangle and a transform.
Further work should expand this to an entire rectangle/transform stack,
but this is an improvement over the current situation.

This was necessary to properly handle clipping in the presence of
nested iframes and `overflow: scroll` stacking contexts, while
simultaneously handling transformed and clipped iframes.

We still have some clipping issues when `overflow: scroll`, iframes, and
transforms are all nested together: this is visible in the browser.html
tab change animation on duckduckgo.com result pages. I wanted to get
this in now, though, because this significantly improves the current
situation.

Addresses servo/servo#11150.
Addresses servo/servo#11151.