Switch the UBO data to be untyped.
This is an interim workaround for how slow shader compilation is on
Linux when using AoS style declarations. We need this change in order
to enable CI testing of WR, otherwise the test pass is too slow.

It brings the shader compile time on Linux down from around 1.8s to
around 0.2 seconds.