Store the last rendered epoch for all pipelines.
Instead of storing the last rendered epoch for each pipeline from
the previous frame, accumulate this information for each frame.
This is a workaround for the way servo's compositor keeps iframes
in the frame tree when they are invisible.

This is a workaround for https://github.com/servo/servo/issues/13149.

Long term, we should find a better solution for this.