Change packed tiles to use vertex textures, and extend to generic render task data.
This is a performance win (by removing all the CPU code that handles batching
tiles) and also implements the framework for adding different types of
render tasks. For example, a prim cache or blur render task may provide
some different data to the shaders than the alpha batcher render task.