Add basic overlay scrollbar support.
This adds the infrastructure for adding scrollbars to any scroll root,
however for now it only draws a scrollbar for the root frame. Once we
are happy with the implementation and design, we can easily enable it
for iframes and other scroll roots.

The idea is to insert a primitive of any kind into the primitive tree
and tag it as being a scroll bar primitive for a given layer. This
means we can allow styling of the scrollbar in the future, making use
of any of the display item features that are supported (e.g. border
corners, gradients, images, anti-aliasing).