Change shaders to compile lazily.
This makes a huge improvement to startup time by compiling shaders
the first time they are used instead of up front.

This improves iteration time during development, but most importantly
it drastically improves the CI time when using webrender, since many
tests only need to compile a small subset of the shaders that are
available.