Add fine grained (per-draw call) GPU profiling.
This provides a graph that shows draw call timing for the
previous 20 frames. The colors identify the type of primitive
batch. The lengths identify the GPU time of each batch, relative
to the maximum frame time of the current graph.

This allows us to see visually:
 - How many draw calls are occurring per frame.
 - Why batches are being split.
 - Which draw calls are being reported as frame spikes.
 - Which shaders are the most expensive for a given scene.
 - How batching changes during scrolling / reflows.
 - How long it takes to execute vertex texture uploads, target switches etc.

This makes it much easier to see the effects of experimenting with
new features and optimizations, such as batcher improvements,
primitive caching etc without having to use apitrace.