Add framework for dynamic/cache render tasks, and use for box shadows.
This adds the framework for tiles and primitives to add dependent
render tasks. The proof of concept in this patch uses them to
speed up box shadow rendering, but the dynamic render tasks can
be used for many things in the future. For example, we can use
this system to render rotated clip masks, or cache text runs
to a texture.

The box shadows are rendered by adding a child task that draws the
upper left corner of a box shadow. The primitive shader for box
shadows then applies this rendered patch over the box shadow by
clamping and mirroring the texture coordinates in the fragment
shader.

Dynamic render tasks are added to passes via a cache key that is
defined based on the type of item being cached. This means that
dynamic render tasks are du-duplicated during allocation and
assignment to render passes and targets. This means that if different
tiles require the same cache primitive, it will only get rendered
once. It also means that if two primitive exist that require the
same resource type (e.g. two box shadows with different colors but
same size and blur radius) then the cache item will only be drawn
once to the child render target.

With this change, GPU time on transparent_rects demo drops from
16ms down to 9ms on my test setup.