Add get_image_properties() to resource cache.
This allows querying image opacity before an image is added to
the texture cache.

This doesn't change functionality, but it's prep work for some
of the upcoming interface changes in the resource cache that are
needed for subpixel glyph functionality.

This is arguably a much cleaner interface as well, it never really
made sense to have this accessible only via the texture cache.