Use glyph metrics to calculate text run bounds and offsets.
Instead of rasterizing glyphs, and using that to calculate the
bounding box of text runs, use glyph metrics instead.

This is important, since it allows us to delay rasterization of
glyphs until later in the frame, which means we can delay the
decision on whether to rasterize a glyph with subpixel
antialising or not.

It also removes the concept of "user data" in the texture cache
since the resource cache is now used to extract glyph metrics.