Fix a bug with clipping when scrolling after a layout.
The layer local rect is clipped against the root viewport
in local space, so it can't be used in a way that it was.

We don't actually need this here anyway. The local clip rect
of the layer is currently handled by write_vertex, and in the
future will be handled when stacking contexts write their
own clip mask.