Export the servo-freetype-sys feature from webrender
While the freetype/servo-freetype-sys feature is desired by servo, it is not
desired by gecko as it needlessly pulls in a dependency on the servo-freetype-sys
library. In order to opt-out of this dependency we have to re-export it from
Webrender. We also need to re-export the codegen feature to allow opting in to
that one explicitly with default-features=false.