Add specific shaders for rectangles with CLIP feature.
Having the base rectangle shader include clip makes the software
mesa implementation (somewhat) slower for reftests. This is enough
of a timing difference that it triggers a lot of intermittent
test failures in Servo.

Instead, include a simple rectangle shader for workloads that
don't have a clip mask.

The other shaders are typically more expensive and not used as
frequently by the reftests, so hopefully this change is enough
to work around the Servo intermittent bugs.