Support dynamic changes to device_pixel_ratio.
Ensures that the device pixel ratio isn't stored in any
structures permanently, and that any code that depends
on the value of the device pixel ratio is moved to sections
that are executed each render.

The idea is that the current device pixel ratio is stored
in the render backend, and this gets passed through the entire
pipeline each frame to the renderer.

In the future there will be an API provided to modify the
device pixel ratio, but for now it is still fixed to a single
value at init time.

This is prep work for supporting functionality such as zoom, and
moving the window between monitors with different device pixel ratios.