Removes the tiling and render task stack from clipping.
This also ensures that primitive clip masks never render
any larger than the primitive bounding rect, which improves
the timings on GitHub a lot.

This does cause a slight performance regression on some
sites (GitHub in particular) but is a first step to the
planned clipping and tiling improvements coming. After those
land, the performance will be better than originally.