Run clip mask shaders only in the region of interest.
For rectangles, this means the clip mask generation runs only on the
four corners. For image masks, the mask generation shader runs only
on the local rect of the image mask.

The rectangle clip mask shader is also much faster now, since it
only needs to consider one corner in the fragment shader.

This drastically reduces the amount of time spent building clip masks.

One remaining issue is that we now allocate large rectangles for large
clip masks. This is a memory waste, but has (almost) no effect on
performance. A follow up PR will use the z-buffer to draw these
large primitives in segments, which will mean the memory allocation
is no longer an issue, and also improve performance of the primitive
shaders (by removing the clip rect check).