Add an optimization for drawing simple borders.
This is primarily to speed up the time it takes to draw
large, simple borders on OSMesa. This makes some tests
run a lot faster when using software rendering, which
reduces the chances of timeouts occurring due to the
large number of tests that the Servo CI machines run
in parallel.

It's also a reasonable optimization for simple borders
when running a hardware renderer.