Fix a regression in text rendering where non-renderable glyphs are present.
The text run code in prepare_prim_for_render() takes account of
non-renderable glyphs and pushs only the renderable glyphs into
the list of glyphs for the text run.

The recent subpixel positioning changes stopped using this and were
using the source glyph indices. This can result in incorrect indexing
of glyph information in the text run shader.

This patch modifies the text run request and render code to use
the glyph indices array. It extends it to contain the x/y coords
of the glyph, allowing the subpixel positioning to work. It also
tidies up some of the code so that the resolve step no longer
requires access to the stacking context and aux maps.