Implement an API that allows animating certain properties of an existing display list.
Currently, this allows animating the transform and opacity of
a stacking context, but it can easily be extended in the future
to handle other property types (e.g. other filters etc).

The current list of property values are passed via the
GenerateFrame message, which ensures that all animated properties
are updated atomically.

The implementation right now is far from optimal. The idea here
is to get the public API correct and implemented, and then we can
focus on optimizing the implementation, once the in-progress work
on scroll layers and reference frames lands.