Fix up the pixel snapping code.
This changes the pixel snapping code to snap to a device pixel
after transformation, rather than snapping in local space.

Previously, fixing this code was difficult to get exactly right
due to the tiling clamp and UV calculation, but now it's very simple.

This adds 31 new passes to the WPT/CSS tests, and 4 new failures. One
of the new failures also fails on Gecko / Blink. The other three new
failures appear to be a bug in osmesa - they don't occur on llvmpipe
or any real GPU hardware I've tested on (on osmesa, a very small number
of pixels in the 3 tests are off by 1/255 in one of the color channels).

Overall, this is a big improvement on the current situation, and means
that we'll be able to start the work to make subpixel positioning correct
in more situations.