Add ClipScrollGroup
ClipScrollGroups represent a portion of a stacking context that is
clipped and scrolled. Before we thought of the entire contents of a
StackingContext as clipped and scrolled in the same way. This isn't the
case on the web, so we add preliminary support for the concept, while
maintaining the 1-to-1 relationship between them for the moment.

Parts of this PR are based on a proof of concept by Glenn Watson.