Properly handle content transformed to empty rect
When content was transformed to an empty rectangle, we didn't always
make its ClipScrollGroup xf_rect member None. This caused an issue when
we tried to use the xf_rect member of the ScrollLayer for masking,
which was properly set to None. We fix this issue by making the
behavior consistent between the two structs.

Fixes #894.