Have ClipScrollTree use reference frame relative coordinates
Instead of assuming that ClipScrollNodes and stacking contexts exist in
a contiguous tree, make the coordinate system of the ClipScrollTree
independent of the one used for stacking contexts. Now all coordinates
are calculated relative to the parent reference frame of every node.

Additionally, stacking contexts no longer have a transformation,
since they already rely on reference frames to be positioned. Stacking
contexts still need a rectangle and an origin though, because display
items may not be positioned relative to the top left corner due to the
way that overflow is calculated in Servo.

This change is necessary in order to clip and transform arbitrary
stacking context items with arbitrary scrolling nodes (within the same
reference frame).