Add support for pushing already built display lists.
Gecko has the concept of 'empty transactions' where we need to get a new
frame composited but aren't rebuilding the Gecko display lists. To
support this we'd like to keep around a retained representation of the
partial WebRender display list.

One easy way to implement this is to just use a BuiltDisplayList
as the retained representation and support appending them
to display lists. Unfortunately, the implementation is a bit fragile
because of needing to deal with ItemRanges.

Hopefully, we can get rid of this eventually and avoid rebuilding the
display list at all in those scenarios.