Assign ScrollLayerId per primitive
Allow assigning a ScrollLayerId directly to a primitive. In the future,
this will allow API users to assign primitive to arbitrary scroll
layers. As part of this work we collapse the scroll_layer_store in
FrameBuilder into the ClipScrollTree itself, making the architecture
a bit less complicated. The biggest potential change here is that
changing between clipping layers requires HashMap lookups instead of
array indexing. If this proves to be a performance issue, it is
possible to optimize this code at the cost of simplicity.