Properly skip invisible stacking contexts
Previously, we tried to skip invisible stacking contexts, but the logic
for that was very broken, because it did not take into account nested
stacking contexts. We fix that issue as well as removing code for
skipping empty scroll layers (which could never work for properly
formed display lists). Finally, we add a test for this behavior.