Add a RadialGradient display item modeled after CSS gradients
A RadialGradient is designed to support the properties needed for a CSS
radial-gradient(). It has a center, a radius X, a radius Y, a collection
of stops, and a repeat flag. This means it can be an ellipse or a
circle. The prim fragment shader does its calculations on a sphere,
so the prim vertex shader sets up a scale transform for ellipses if
needed.