Stop clipping on stacking context boundaries
Now that clipping regions are first-class elements in the display list,
it doesn't make as much sense to clip along stacking context
boundaries. Additionally, this removes a lot of complexity around the
difference between stacking context bounds and overflow.

Using the transparent rectangle test, I couldn't detect a difference in
performance, and that creates a lot of stacking contexts.