Normalize gradient stops
Gradients can be defined with stops outside the range of [0, 1].
When this happens the gradient needs to be normalized by adjusting
the gradient stops and gradient line into an equivalent gradient
with stops in the range [0, 1].

This allows for proper repeating behavior for gradients because the
range from [first_stop, last_stop] must be repeated, not the range
[0, 1]. This also allows for proper rendering of gradients with stops
beyond the range [0, 1] because the GradientData struct only works
for stops in the range [0, 1].

LinearGradients and RadialGradients are updated to be normalized by
default. The DisplayListBuilder method for RadialGradient is updated
to be simpler and more in line with the CSS representation of a
radial-gradient(). The old method is kept for possible SVG uses in the
future, and replay capabilities.