Don't use an aligned gradient when the gradient line doesn't fill rect
The aligned gradient fast path doesn't correctly fill all of its display
item's rect if it's gradient line doesn't intersect the whole rect. This
is because extra geometry is not added to fill in that region. This can
be fixed, but for correctness now we should use the angle gradient
shader for these cases.