Allow tiling of radial and linear gradients
This commit adds `tile_size` and `tile_spacing` fields to linear and
radial gradients. Tiles start at `rect.origin` and are `tile_size` with
`tile_spacing` between each tile.

Tiling is implemented inside the gradient shaders, not by decomposing
the display item into multiple primitives. This has the advantage of
being simple to implement, and in the best case where there is no
spacing it should be just as fast as transforming the gradient into a
repeating gradient.

This has the disadvantage that large tile spacing can cause many
dropped fragments. There may be good reason to investigate decomposing
to multiple primitives in these cases.