Fix the way that clip positioning is specified
Make the way that clip positions are specified a bit more predictable.
Now the display item bounds are the "content bounds," which specify the
location and size of the content that goes inside this clip. The clip
coordinates are specified relative to this location, while the content
size is only used for scrolling.

Currently we only support clips rectangles that start at the origin of
the content, which is equivalent to the capabilities we had before.
Later we can add this functionality, if it is necessary.

Also, add a test for this as well as updating all reftest files to
reflect the new arguments.