Fix aligned gradient condition
We can't use the aligned gradient shader when the gradient start
and end do not cover the entire portion of rect. In this case, the
aligned gradient path will not cover the entire rect causing an
incorrect rendering. The condition added to fix this was flipped
backwards, causing the aligned gradient shader to never be used.