Interpret gradient start and end positions as relative to bounds
Gradient start and end positions are currently relative to their
parent stacking context. This isn't necessary and complicates the
representation of tiled gradients.

This commits reinterprets the start and end point of gradients as
specified relative to their display item bounds. This affects
GradientDisplayItem, RadialGradientDisplayItem and BorderDisplayItem.

Two border gradient reftest ref files are fixed because they relied on
the old behavior for specifying the start and stop position. The rest of
the reftests do not need to be changed, as they used stacking contexts.