Fix combined_local_viewport_rect for nested nodes
The calculation of the combined_local_viewport_rect was being done
using the accumulated_scroll_offset, but this was not correct, because
the parent combined_local_viewport_rect was already calculated to be
independent of those offsets. This means that the only scrolling offset
that matters is one between the parent and the child.

Add a test for this as well.