Clear render targets with rgba(0,0,0,0) and remove isolate_clear
The composite shader relies on its render target being premultiplied.
This causes issues with transparent surfaces behind a mix-blend-mode.
The clipping code no longer relies on the previous behavior, so this
can be changed.

The isolate_clear code then no longer has an effect, so that is removed
too. Notably, the code for isolating stacking context's remains as only
the code that clears the rects is no longer needed.