Make image tiling work with external buffer images.
Tiling is already handled in the resource cache by specifying the proper size, stride and offset in order to upload the correct part of the image (the tile) to the texture atlas. Getting this to work with external buffer images is only a matter of passing the correct offset which was ignored before this commit.