Make the BlobImageRenderer manage the blob objects.
Before this commit BlobImageRenderer implementations are not expected to preserve state between frames as the resource cache are retaining the blobs and passing them as paramters of the blob requests. This approach is simple but forces blob image updates to contain the whole blob and forces the blob renderer to perform deserialization each time. With this change the blob renderer is responsible for storing the drawing commands (in whichever form) and has more freedom in the way it handles updates.