Handle RadialGradients with zero sized radii
A RadialGradient with radius of zero should only display the final color
stop value.

Currently, a radius of zero will cause ratio_xy to be NaN, Inf, or -Inf and
the output will be blank. Forcing the radius to zero and ratio_xy to one
won't work because ps_radial_gradient can't handle a zero sized radius. So
instead we can construct an equivalent gradient with only the last color stop.

Alternatively this could be caught on the backend and the RadialGradient
could be converted to a color rect, but that would disable dithering.
This is a simple enough fix and makes reftests consistent.