Add a feature for dithering
In order to help pass all the reftests for enabling gradients by default in
Gecko, it would be nice to be able to disable dithering. Canvas2D gradients
and border gradients in Gecko don't use dithering, and that causes some
issues. Enabling dithering can be investigated as a follow up.