Resolve the ext-image again even though it's resolved in previous frame.
Maybe we will do the partial updating for the frame data(scrolling?).
So, the SourceTexture in cpu_yuv_image is not in a invalid status. Then,
we will skip that ext-image resolving. Thus, it will not be put into the
deferred_resolves list. Finally, we can't resolve that ext-image in
renderer.