Refactor multi-threaded glyph rasterization.
This commit replaces the glyph rasterization code in resource_cache.rs with a new and separate implementation. The new glyph rasterizer uses rayon's thread pool with a dedicated thread and without the need to move the glyph cache between threads. Adding/removing font templates is simplified and does not block the entire thread pool anymore.