Premultiply gradient stop colors
CSS gradient stop colors are interpolated in premultiplied space.
Currently we assume that color stops that are passed in are
premultiplied, but because it's needed in both Servo and Gecko, it makes
sense to do the premultiplication in WebRender and expose a straight
alpha color like the rest of the API does.

The premultiplication is purposefully not done in DisplayListBuilder
to prevent it from being saved for replay and then premultiplied again
when we replay the display list.