Hold old resources in the texture cache for a bit longer.
Previously, we were expiring resources from the texture cache
as long as they weren't used that frame. This was resulting in
a lot of extra glyph rasterization than we really want.

Instead, only expire resources that haven't been used for ~10 seconds
when running at 60fps. This drastically reduces the time worker
threads spend rasterizing glyphs, and doesn't have any noticeable
effect on the size of the texture cache from the sites I tested on.